Products adding to cart without reloading page

+++ database/migrations/2017_05_18_040114_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->integer('category_id')->nullable()->index()->unsigned();
             $table->string('title');
             $table->string('description');
-            $table->decimal('price', 5, 2)->default(0);
+            $table->integer('price')->default(0);
             $table->integer('brand_id')->unsigned()->nullable();
             $table->integer('rating')->unsigned()->nullable();
             $table->integer('stock')->unsigned()->nullable()->default(0);
